fix(#1539): fix mmCalculator input and output formatting
