+++ src/filtertransdialog.h
@@ -25,7 +25,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "reports/htmlbuilder.h"
 
 #include <wx/dialog.h>
-#include "mmtextctrl.h"
+#include "mmTextCtrl.h"
 
 
 class mmFilterTransactionsDialog: public wxDialog
