+++ src/maincurrencydialog.cpp
@@ -23,9 +23,8 @@
 #include "constants.h"
 #include "defs.h"
 #include "images_list.h"
-#include "mmCalculator.h"
+#include "mmTextCtrl.h"
 #include "mmSimpleDialogs.h"
-#include "mmtextctrl.h"
 #include "paths.h"
 #include "util.h"
 #include "validators.h"
