+++ src/model/Model_Currency.h
@@ -77,7 +77,7 @@ public:
     static wxString toStringNoFormatting(double value, const Data* currency = GetBaseCurrency(), int precision = -1);
     /** convert value to a currency formatted string with required precision */
     static wxString toString(double value, const Data* currency = GetBaseCurrency(), int precision = -1);
-    /** Reset currency string like 1.234,56 to standard number format like 1234.56 */
+    /** Reset currency string like 1.234,56 to standard C locale number format like 1234.56 */
     static const wxString fromString2Default(const wxString &s, const Data* currency = Model_Currency::GetBaseCurrency());
     static bool fromString(wxString s, double& val, const Data* currency = GetBaseCurrency());
     static int precision(const Data* r);
