+++ src/usertransactionpanel.h
@@ -22,7 +22,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "constants.h"
 #include "defs.h"
 #include <wx/spinbutt.h>
-#include "mmtextctrl.h"
+#include "mmTextCtrl.h"
 #include "model/Model_Translink.h"
 
 class Model_Currency;
